reorganize host initialization
Due to complaints of spaghetti code, and difficult to follow
initialization, the initialization has been restructured in an attempt
to make things more clear.
